requirements.txt
Django>=2.2.13,<3
psycopg2==2.8.5
#for MySQL: MySQL-python==1.2.5
statsd==3.3.0
pep8==1.7.1
pyflakes==2.2.0
mccabe==0.6.1
pycodestyle==2.6.0
configparser==5.0.0
importlib-metadata==1.6.1  # for flake8
zipp==3.1.0
flake8==3.8.3
lxml==4.5.1
extras==1.0.0
fixtures==3.0.0
pbr>=0.11
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1  # django-debug-toolbar
python-mimeparse==1.6.0  # tastypie
python-dateutil==2.8.1  # tastypie
defusedxml==0.6.0  # tastypie
django-tastypie==0.14.2 # pyup: < 0.14.3
coverage==5.1
logilab-common==1.7.1
lazy-object-proxy==1.5.0
wrapt==1.12.1
unicodecsv==0.14.1

editdistance==0.5.3

pylint==2.5.3

astroid==2.4.2

pyasn1==0.4.8
ndg-httpsclient==0.5.1

chardet==3.0.4

# https://urllib3.readthedocs.io/en/latest/user-guide.html#ssl-py2
idna==2.9
urllib3==1.25.9  # sentry-sdk
certifi==2020.6.20  # sentry-sdk
cryptography==2.9.2
pyOpenSSL==19.1.0
asn1crypto==1.3.0  # for pyOpenSSL
cffi==1.14.0  # for pyOpenSSL

requests==2.24.0
contextlib2==0.6.0.post1
rcssmin==1.0.6
rjsmin==1.1.0

djangowind==1.2.3
django-tagging==0.5.0
django-reversion==3.0.7

requirements/src/djangohelpers-0.20-py2.py3-none-any.whl
django-contrib-comments==1.9.2
django-threadedcomments==1.2
django-courseaffils==2.2.4
django-statsd-mozilla==0.4.0
sentry-sdk==0.15.1
django-appconf==1.0.4
django-compressor==2.4
django-stagingcontext==0.1.0
django-ga-context==0.1.0
django-smoketest==1.1.2

django-debug-toolbar==2.2

django-markwhat==1.6.2
ipaddress==1.0.23
text-unidecode==1.3  # for faker
Faker==4.1.1
factory_boy==2.12.0
django-impersonate==1.5.1
django-registration-redux==2.8
django-waffle==1.0.0
django-cors-headers==3.4.0

# Django >= 3.1 has full support for samesite cookies. 
# Until then, we'll need to use this.
django-cookies-samesite==0.6.3

httplib2==0.18.1
oauth==1.0.1
oauth2==1.9.0.post1
oauthlib==3.1.0
pylti==0.7.0
nameparser==1.0.6
django-bootstrap3==14.0.0
ua_parser==0.10.0
user_agents==2.1
django_user_agents==0.4.0
gunicorn==20.0.4
path==14.0.1
simplegeneric==0.8.1
scandir==1.10.0
pathlib2==2.3.5

freezegun==0.3.15
django-smtp-ssl==1.0
djangorestframework==3.9.4  # pyup: <3.10

ccnmtlsettings==1.8.3
django-extensions==2.2.9

s3transfer==0.3.3
jmespath==0.10.0

botocore>=1.16.0,<1.18.0
boto3>=1.13.0,<1.15.0

entrypoints==0.3
django-storages==1.9.1
django-cacheds3storage==0.2.1

requests-toolbelt==0.9.1
pytz==2020.1
isodate==0.6.0
cached-property==1.5.1

appdirs==1.4.4  # zeep
attrs==19.3.0  # zeep
zeep==3.4.0

stevedore==2.0.1
pyyaml==5.3.1
bandit==1.6.2
funcsigs==1.0.2 
pypanopto==0.2.0
entrypoints==0.3

billiard==3.6.3.0
anyjson==0.3.3
vine==1.3.0
amqp==2.6.0 # pyup: < 2.5
amqplib==1.0.2
kombu>=3.0.37,<4.0.0
celery==3.1.26.post2 # pyup: <4.0.0
django-celery>=3.2.2,<3.3.0
# memcached
pylibmc==1.6.1; sys_platform == 'linux'
